Update tc-print to disassemble Aarch64 with inspection comments
