+++ hphp/tools/tc-print/offline-ppc64-code.cpp
@@ -1,0 +1,65 @@
+/*
+   +----------------------------------------------------------------------+
+   | HipHop for PHP                                                       |
+   +----------------------------------------------------------------------+
+   | Copyright (c) 2010-present Facebook, Inc. (http://www.facebook.com)  |
+   +----------------------------------------------------------------------+
+   | This source file is subject to version 3.01 of the PHP license,      |
+   | that is bundled with this package in the file LICENSE, and is        |
+   | available through the world-wide-web at the following url:           |
+   | http://www.php.net/license/3_01.txt                                  |
+   | If you did not receive a copy of the PHP license and are unable to   |
+   | obtain it through the world-wide-web, please send a note to          |
+   | license@php.net so we can mail you a copy immediately.               |
+   +----------------------------------------------------------------------+
+*/
+#include "hphp/tools/tc-print/offline-code.h"
+
+#include <stdio.h>
+#include <cxxabi.h>
+#include <vector>
+#include <assert.h>
+#include <iomanip>
+#include <sys/stat.h>
+
+#include "hphp/tools/tc-print/tc-print.h"
+#include "hphp/tools/tc-print/offline-trans-data.h"
+
+#include "hphp/util/disasm.h"
+#include "hphp/vixl/a64/disasm-a64.h"
+#include "vixl/a64/instructions-a64.h"
+
+using std::vector;
+
+namespace HPHP { namespace jit {
+
+#if defined(__powerpc64__)
+
+TCA OfflineCode::collectJmpTargets(FILE  *file,
+                                      TCA    fileStartAddr,
+                                      TCA    codeStartAddr,
+                                      uint64_t codeLen,
+                                      vector<TCA> *jmpTargets) {
+
+// TODO:
+}
+
+// Disassemble the code from the given raw file, whose initial address is given
+// by fileStartAddr, for the address range given by
+// [codeStartAddr, codeStartAddr + codeLen)
+
+void OfflineCode::disasm(FILE* file,
+                            TCA   fileStartAddr,
+                            TCA   codeStartAddr,
+                            uint64_t codeLen,
+                            const PerfEventsMap<TCA>& perfEvents,
+                            BCMappingInfo bcMappingInfo,
+                            bool printAddr,
+                            bool printBinary) {
+
+// TODO:
+}
+
+#endif
+
+} } // HPHP::jit
